* use DTO classes for id/uuid wrappers in UnityCatalogService undocumented APIs * move DTO classes into unity package
