+++ src/main/java/com/edmunds/rest/databricks/DTO/unity/MetastoreDTO.java
@@ -1,4 +1,4 @@
-package com.edmunds.rest.databricks.DTO.account;
+package com.edmunds.rest.databricks.DTO.unity;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Getter;
